refactor: Command 모듈 Usecase 원래대로 패키지로 변경
