+++ domain/mysql/src/main/kotlin/com/hs/repository/ProductQueryRepository.kt
@@ -1,9 +1,7 @@
 package com.hs.repository
 
 import com.hs.dto.FindProductDto
-import com.hs.entity.Product
 
 interface ProductQueryRepository {
     fun findProductAggregate(id: Long): FindProductDto?
-    fun findProduct(id: Long): Product?
 }
